requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy>=1.26.0
opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
tqdm>=4.66.0
mysqlclient>=2.2.0
python-multipart>=0.0.6
